lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
psycopg2-binary==2.9.5
pytz==2022.6
python-dotenv==0.21.0